+++ src/app/modules/student-answers/state/student-answers.state.ts
@@ -5,10 +5,14 @@ export interface StudentAnswersState {
 	studentAnswerShown: StudentAnswer | null;
 	studentAnswerShownStatus: ResourceStatus;
 	sendingProcessTextRequest: boolean;
+	studentAnswersShown: StudentAnswer[] | null;
+	studentAnswersShownStatus: ResourceStatus;
 }
 
 export const studentAnswersInitialState: StudentAnswersState = {
 	studentAnswerShown: null,
 	studentAnswerShownStatus: ResourceStatus.NotLoaded,
-	sendingProcessTextRequest: false
+	sendingProcessTextRequest: false,
+	studentAnswersShown: null,
+	studentAnswersShownStatus: ResourceStatus.NotLoaded
 };